refactor: move `Course` and `Visibility` to model
