+++ app/src/androidTest/java/com/github/onlynotesswent/ui/overview/OverviewTest.kt
@@ -19,8 +19,8 @@ import com.github.onlynotesswent.model.users.UserRepository
 import com.github.onlynotesswent.model.users.UserViewModel
 import com.github.onlynotesswent.ui.navigation.NavigationActions
 import com.github.onlynotesswent.ui.navigation.Screen
-import com.github.onlynotesswent.utils.Course
-import com.github.onlynotesswent.utils.Visibility
+import com.github.onlynotesswent.model.common.Course
+import com.github.onlynotesswent.model.common.Visibility
 import com.google.firebase.Timestamp
 import org.junit.Before
 import org.junit.Rule
