+++ app/src/main/java/com/github/onlynotesswent/model/folder/FolderRepositoryFirestore.kt
@@ -1,7 +1,7 @@
 package com.github.onlynotesswent.model.folder
 
 import android.util.Log
-import com.github.onlynotesswent.utils.Visibility
+import com.github.onlynotesswent.model.common.Visibility
 import com.google.firebase.Firebase
 import com.google.firebase.auth.auth
 import com.google.firebase.firestore.DocumentSnapshot
