+++ app/src/test/java/com/github/onlynotesswent/utils/NotesToFlashcardTest.kt
@@ -1,6 +1,8 @@
 package com.github.onlynotesswent.utils
 
 import android.content.Context
+import com.github.onlynotesswent.model.common.Course
+import com.github.onlynotesswent.model.common.Visibility
 import com.github.onlynotesswent.model.file.FileType
 import com.github.onlynotesswent.model.file.FileViewModel
 import com.github.onlynotesswent.model.flashcard.Flashcard
